move keybinding api into more navigable structure
