+++ src/main/java/io/ejekta/kambrik/Kambrik.kt
@@ -6,6 +6,7 @@ import io.ejekta.kambrik.logging.KambrikLoggingApi
 import io.ejekta.kambrik.message.KambrikMessageApi
 import io.ejekta.kambrik.serial.KambrikSerialApi
 import io.ejekta.kambrik.structure.KambrikStructureApi
+import io.ejekta.kambrikx.input.KambrikInputApi
 import org.apache.logging.log4j.LogManager
 
 
@@ -22,6 +23,8 @@ object Kambrik {
     val Message = KambrikMessageApi()
 
     val Serial = KambrikSerialApi()
+
+    val Input = KambrikInputApi()
 
     internal val Logger = LogManager.getLogger("Kambrik")
